+++ pkg/store/blueprint/currencies.go
@@ -2,7 +2,7 @@ package blueprint
 
 import (
 	"fmt"
-	"github.com/voyago/converter/src/support"
+	"github.com/voyago/converter/pkg/support"
 	"runtime"
 	"strings"
 )
@@ -24,7 +24,7 @@ func MakeCurrenciesBlueprint() (Blueprint, error) {
 
 func (body Blueprint) SourcePath() string {
 	_, fileName, _, _ := runtime.Caller(0)
-	baseDir := strings.Split(fileName, "/src/store/blueprint/currencies.go")[0]
+	baseDir := strings.Split(fileName, "/pkg/store/blueprint/currencies.go")[0]
 
 	fmt.Println("===============>", baseDir)
