+++ pkg/store/handler/handler.go
@@ -1,6 +1,6 @@
 package handler
 
-import "github.com/voyago/converter/src/model"
+import "github.com/voyago/converter/pkg/model"
 
 type Handler interface {
 	ExchangeRates() (model.Currencies, error)
