+++ pkg/support/strings.go
