+++ tests/unit/conversion/converter_test.go
@@ -1,8 +1,8 @@
 package conversion
 
 import (
-	"github.com/voyago/converter/src/conversion"
-	"github.com/voyago/converter/src/model"
+	"github.com/voyago/converter/pkg/conversion"
+	"github.com/voyago/converter/pkg/model"
 	"github.com/voyago/converter/tests/mock"
 	"testing"
 )
